==> 03_code/01b_sentence_segmentation.ipynb <==
import spacy
import pandas as pd
import xx_sent_ud_sm

nlp = spacy.load("xx_sent_ud_sm")
nlp = xx_sent_ud_sm.load()


data = pd.read_csv("../04_clean_data/missing_speeches_parlee.csv")

# Step 1: Clean the DataFrame
data['text'] = data['text'].astype(str)

# Step 2: Feed the DataFrame to spaCy pipeline
docs = list(nlp.pipe(data.text, n_process=4))

# Step 3: Create a new DataFrame with sentences
sentences = []
doc_ids = []
sentence_ids = []

for doc_id, doc in zip(data.text_id, docs):
    sent_id = 1
    for sent in doc.sents:
        sentences.append(sent.text)
        doc_ids.append(doc_id)
        sentence_ids.append(sent_id)
        sent_id += 1

result_df = pd.DataFrame({
    'text_id': doc_ids,
    'Sentence_id': sentence_ids,
    'sentence': sentences
})

result_df

result_df.head(40)

data[["text_id","session_id","id_speaker"]].head(5)

data.drop(columns=['text'], inplace=True)

# Left join the 'data' DataFrame onto the 'result_df' DataFrame using the 'ID' column
merged_df = pd.merge(data[["text_id","session_id","id_speaker"]], result_df, left_on='text_id', right_on='text_id', how='left')

merged_df.to_csv("../04_clean_data/missing_speeches_parlee_sents.csv")
